Sounds, Scripts, Animations Added
Created folders for everyone's separates audio files so we can help organise them. Added a scripts and animation folder and created two scripts to randomise timing for audio. Moved location of enemy to behind the wall in sight and added audio to him. Added animation and audio to starting door, added audio and scripts to lights and chains. Added ambience soundtrack to player. Created and added 3D sound of wind and water in open areas.